requirements.txt
torch
torchvision
numpy
pyyaml
pillow 